services/brain-python/requirements.txt
pyzmq==26.2.0
protobuf==5.29.0
grpcio-tools==1.68.1
pyyaml==6.0.2
